Dockerfile Setup
I can now build an image and run a container off of the frontend.Dockerfile! There's an issue with sending a SIGINT to the storybook server that I need to figure out.  For the time being, cd into the react_app folder, build the image with "docker build -t frontend-image -f frontend.Dockerfile . --debug", and run the container with "docker run -p 5173:5173 -p 6006:6006 frontend-image". You can command + click the localhost URLs to access both pages. The Docker compose file is not detecting Vite for some reason.