Homepage and Learning Paths Components
The homepage is structurally complete, and the learning path components are almost structurally complete at this point aside from the learning paths accordion.  Next steps are to finish off the accordion and get the profile structurally complete. Then I move on to theming and styling and getting everything into storybook.