Updated Colors, Added Many Storybook Components
Warnings for dark mode colors have been resolved. Need to add the rest of the components to Storybook and double check for functionality and style adherence. Font size is incorrect in many places atm, and the warnings about Geist not being found and having a random incorrect calc statement also need to be resolved.