Updated Banner and FocusWrapper
Made sure the spacing on the banner component was consistent and that the FocusWrapper could scroll.

TODO:
Need to fix the eventTime and image in the Banner component. Need to add the pagination component to the components on the homepage. Items inside the FocusWrapper should overflow when scrolling instead of being cut off.