Added More Interfaces, Updated Learning Path Overview, Content, and Pages
Added interfaces needed for the challenges, resources, and learning path information. Updated LearningPathOverview to LearningPathOverviewContent. Currently working on switching the content shown on the left side of the page based upon what is selected on the right side of the page (or the previous/next buttons that will be on the bottom left).