Updated Learning Path Flow, ResourceGroup Div, Challenge Div
Changed skeletons in the learning path detail page so that they loaded properly. Updated the hint accordion and reformatted the challenge div slightly. Updated the challenge mock data to include hints and tags. Need to update the ResourceGroup, Game, and Challenge div components to take in IDs rather than the actual data, and make the request for the data within each respective component.